// src/app/views/layout/excel-sheets/optionList.interface.ts
export interface OptionList {
    id: number;
    optionName: string;
    isCorrect: boolean;
    createdAt: string;
    questionId: number;
    questionName: string;
    questionDegree: number;
    examId: number;
    examName: string;
    examShortDescription: string;
    examDescription: string;
}